Crate rispc

A library for build scripts to compile Intel SPMD (ispc) code.

This library is intended to be used as a build-dependencies entry in Cargo.toml, with a runtime component added to the dependencies:

[build-dependencies]
rispc = "0.1"

[dependencies]
rispcrt = "0.1"

The purpose of this crate is to provide the utility functions necesasry to compile ispc code into a static archive which is then linked into a Rust crate. The top-level compile_library function serves as a convenience and more advanced configuration is available through the Config builder.

This crate will automatically detect situations such as cross compilation or other environment variables set by Cargo and will build code appropriately.

Examples

Use the default configuration:

// build.rs
extern crate rispc;

fn main() {
  rispc::compile_library("libmandelbrot.a", &[ "src/mandelbrot.ispc" ]);
}

Use more advanced configuration:

// build.rs
extern crate rispc;

fn main() {
  rispc::Config::new()
    .file("src/mandelbrot.ispc")
    .define("FOO", Some("bar"))
    .math_lib(rispc::Math::Fast)
    .enable_fast_math(true)
    .addressing(rispc::Addr::A64)
    .compile("libmandelbrot.a");
}

For a complete, working example, see the rispc-demo folder in the repository.

Structs

Config
Extra configuration to pass to ispc.

Enums

Addr
An addressing scheme. By default, ispc uses 32-bit addressing. If your Arrays grow to more than 2^32 elements, this will need to be changed to 64-bit.

Arch
The architecture to target. This will generally be autodetected from Cargo's current target, but may be changed manually to either x86 or x86_64 systems.

Cpu
The CPU families one may generate code for. This will generally be autodetected based on the current set of ispc targets, but may be manually overridden.

Math
Selects which math libraries to call out to.

Target
Selects which target ISA(s) and the lane width(s) to generate code for.

Functions

compile_library
Compile a library from the given set of input .ispc files.