Struct rispc::Config

pub struct Config {
    // some fields omitted
}

Extra configuration to pass to ispc.

Methods

impl Config

fn new() -> Config

Constructs a new instance of a blank set of configuration.

The builder is finished with the compile function.

fn addressing(&mut self, a: Addr) -> &mut Self

Sets the addressing mode of the compiled ispc code.

By default, all ispc-generated code is addressed with i32s, so if your arrays have more than 2 billion elements, you'll need to change the addressing mode to Addr::X86_64.

Default value: Addr::X86

fn cpu(&mut self, c: Cpu) -> &mut Self

Adds an entry to the target CPU set.

Code will be generated for all CPUs in the target CPU set, and the correct version to run will be autodetected at run time.

By default, the target CPU set is empty, as it can be automaatically determined by the selected targets.

Default value: []

fn define(&mut self, k: &str, v: Option<&str>) -> &mut Self

Specifies a -D variable with an optional value.

Default value: []

fn debug(&mut self, val: bool) -> &mut Self

Turns on or off generation of debug info.

This will generally be automatically determined by the currently selected cargo profile.

fn math_lib(&mut self, m: Math) -> &mut Self

Selects the math library to call out to.

Default value: Math::Default

fn file<P: AsRef<Path>>(&mut self, p: P) -> &mut Self

Adds a file to the set of files to be compiled together.

Default value: []

fn opt_level(&mut self, level: u32) -> &mut Self

Set the optimization level.

Default value: inferred from current cargo profile

fn enable_assertations(&mut self, val: bool) -> &mut Self

Enables or disables asssertations in the code.

Default value: true

fn enable_fma(&mut self, val: bool) -> &mut Self

Enables or disables generation of fused multiply-add instructions.

Default value: true

fn enable_loop_unroll(&mut self, val: bool) -> &mut Self

Enables or disables loop unrolling.

Default value: true

fn enable_fast_masked_vload(&mut self, val: bool) -> &mut Self

Enables or disables faster masked vector loads on SSE. This may cause reads to go off the end of arrays.

Default value: false

fn enable_fast_math(&mut self, val: bool) -> &mut Self

Enables or disables non-IEEE-754 compliant math operations.

Enabling this option may cause code to go much faster, but have unpredictable error in the result, and corner cases such as inf and NaN may not be handled properly.

Default value: false

fn force_aligned_memory(&mut self, val: bool) -> &mut Self

Enables or disables the generation of aligned vector load and store instructions.

Enabling this option may make code run faster, but it will result in undefined behavior if buffers are not aligned correctly.

Default value: false

fn pic(&mut self, val: bool) -> &mut Self

Enables or disables the generation of position-independent code.

This should generally be true, unless you have a really good reason otherwise.

Default value: true on x86_64, false on x86.

fn target(&mut self, t: Target) -> &mut Self

Adds a target to generate code for.

The first target that's added replaces the default list. Every subsequent target will be added to this new list of targets. Only one target per ISA may be selected. For example, you may have the simultaneous choises of [ Sse2_i32x4, Sse4_i32x8 ], but not [ Sse2_i32x4, Sse2_i32x8 ].

Default value: [ Sse2, Sse4, Avx1, Avx1_1, Avx2 ]

fn werror(&mut self, val: bool) -> &mut Self

Force all warnings as errors.

If enabled, warnings will break the build. If there are warnings, not being treated as errors, cargo will currently silently eat them. Sorry.

Bikeshed about it here.

Default value: true

fn warn(&mut self, val: bool) -> &mut Self

Enables or disables warnings.

Default value: true

fn warn_perf(&mut self, val: bool) -> &mut Self

Enables or disables warnings about suboptimal code performance.

These will not turn into errors even with .werror(true), so cargo will silently ignore them anyhow. This is really suboptimal. Sorry.

Bikeshed about it here.

Default value: true

fn compile(&self, output: &str)

Runs the compiler, generating the output.

The name output must begin with lib and end with .a.